fix the data type for faster and reduction of memory consumption in basic functions in parser.jl

+++ src/utils.jl
@@ -1,12 +1,12 @@
 function condition_set()
     print("GloVe(0) or random(1): ")
-    embedding_type = parse(Int,readline())
+    embedding_type = parse(UInt16,readline())
     if embedding_type != 0 && embedding_type != 1
         print("Error: embedding type")
         exit()
     end
     print("embedding dim: ")
-    embedding_dim = parse(Int,readline())
+    embedding_dim = parse(UInt16,readline())
     if embedding_type == 0
         if embedding_dim in [50,100,300]
             embedding_type = "GloVe"
@@ -28,10 +28,10 @@ end
 function load_content_vocab(path_to_content_vocab::String)
     f = open(path_to_content_vocab)
     info_list = readlines(f)
-    content_vocab = Dict{String,Int}()
+    content_vocab = Dict{String,UInt16}()
     for info in info_list
         info = split(info)
-        content_vocab[info[1]] = parse(Int,info[2])
+        content_vocab[info[1]] = parse(UInt16,info[2])
     end 
     return content_vocab
 end
@@ -39,13 +39,13 @@ end
 function load_binary_rule(path_to_binary_rule::String)
     f = open(path_to_binary_rule)
     info_list = readlines(f)
-    num_category = parse(Int,info_list[1])
+    num_category = parse(UInt16,info_list[1])
     binary_rule = zeros(Bool,num_category,num_category,num_category)
     for info in info_list[2:end]
         info = split(info)
-        left = parse(Int,info[1])
-        right = parse(Int,info[2])
-        parent = parse(Int,info[3])
+        left = parse(UInt16,info[1])
+        right = parse(UInt16,info[2])
+        parent = parse(UInt16,info[3])
         binary_rule[left,right,parent] = 1
     end
     return binary_rule
@@ -54,12 +54,12 @@ end
 function load_unary_rule(path_to_unary_rule::String)
     f = open(path_to_unary_rule)
     info_list = readlines(f)
-    num_category = parse(Int,info_list[1])
+    num_category = parse(UInt16,info_list[1])
     unary_rule = zeros(Bool,num_category,num_category)
     for info in info_list[2:end]
         info = split(info)
-        child = parse(Int,info[1])
-        parent = parse(Int,info[2])
+        child = parse(UInt16,info[1])
+        parent = parse(UInt16,info[2])
         unary_rule[child,parent] = 1
     end
     return unary_rule
@@ -70,5 +70,3 @@ function load_sentence_list(path_to_raw_sentence::String)
     sentence_list = readlines(f)
     return sentence_list
 end
-
-
